requirements.txt
flask 
flask-sock
python-dotenv
openai
google-cloud-speech
gunicorn